some more modules renaming
